fix(production seed): insert roles and sni areas
